refactor: 함수 분리(유틸, 서비스)
